benchmark/geospatial-workflow/requirements.txt
affine==2.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
branca==0.3.1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
descartes==1.1.0
diskcache==4.1.0
docutils==0.15.2
entrypoints==0.3
enum34==1.1.6
Fiona==1.8.13
geojson==2.5.0
geomet==0.2.1
glob2==0.7
grass-session==0.1
html2text==2019.9.26
ibm-cos-sdk==2.5.5
ibm-cos-sdk-core==2.5.5
ibm-cos-sdk-s3transfer==2.5.5
idna==2.8
importlib-metadata==1.2.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.2.0
kiwisolver==1.1.0
line-profiler==3.0.2
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.1.2
memory-profiler==0.55.0
mercantile==1.1.2
mistune==0.8.4
more-itertools==8.0.2
munch==2.5.0
nbconvert==5.6.1
nbformat==4.4.0
numexpr==2.7.0
numpy==1.17.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psutil==5.6.7
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.5.2
pyparsing==2.4.5
pyproj==2.4.2.post1
pyrsistent==0.15.6
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
rasterio==1.1.1
rasterstats==0.14.0
requests==2.22.0
scipy==1.3.3
sentinelsat==0.13
Shapely==1.6.4.post2
tblib==1.6.0
tornado==6.0.3
tqdm==4.40.2
traitlets==4.3.3
traittypes==0.2.1
wcwidth==0.1.7
xarray==0.14.1
zipp==0.6.0
